+++ src/extra.rs
@@ -2,7 +2,7 @@
 
 #[cfg(feature = "bevy_state")]
 pub mod bevy_state;
-#[cfg(feature = "entity_scope")]
-pub mod entity_scope;
+#[cfg(feature = "react")]
+pub mod react;
 #[cfg(feature = "split")]
 pub mod split;